v1.3.0 Description as an attribute
Description being a property, as before, is problematic because it requires the application to instanciate each class just to know the description.
This was acceptable in the past, because it was still a light operation, however, since dependency injection was added, this can potentially cause all sorts of problems.
Backwards compatability is enabled for older applications made before v1.3.0, which still uses instantiations.